requirements.txt
transformers>=4.40.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
einops>=0.6.0
pandas>=1.5.0
rich>=13.0.0
psutil>=5.9.0